Move hosting.h to include/rocksdb
